# refactor: eliminate TestSingletonManager - Phase 1.3 complete (25.1% performance gain)