--- a/scripts/generate-performance-chart.js
+++ b/scripts/generate-performance-chart.js
@@ -6,95 +6,127 @@
  */
 
 const colors = {
-  reset: '\x1b[0m',
-  bright: '\x1b[1m',
-  green: '\x1b[32m',
-  yellow: '\x1b[33m',
-  blue: '\x1b[34m',
-  red: '\x1b[31m',
-  cyan: '\x1b[36m',
-  magenta: '\x1b[35m'
+  reset: "\x1b[0m",
+  bright: "\x1b[1m",
+  green: "\x1b[32m",
+  yellow: "\x1b[33m",
+  blue: "\x1b[34m",
+  red: "\x1b[31m",
+  cyan: "\x1b[36m",
+  magenta: "\x1b[35m",
 };
 
 function generateChart() {
-  console.log(`\n${colors.bright}${colors.cyan}TEST INFRASTRUCTURE SIMPLIFICATION PROGRESS${colors.reset}\n`);
-  
+  console.log(
+    `\n${colors.bright}${colors.cyan}TEST INFRASTRUCTURE SIMPLIFICATION PROGRESS${colors.reset}\n`,
+  );
+
   // Performance Timeline
   console.log(`${colors.bright}Performance Timeline:${colors.reset}`);
-  console.log('┌─────────────┬──────────────┬──────────────┬─────────────┐');
-  console.log('│   Baseline  │   Phase 1.1  │   Phase 1.3  │  Phase 2    │');
-  console.log('│   (Start)   │  (Complete)  │  (Complete)  │  (Planned)  │');
-  console.log('├─────────────┼──────────────┼──────────────┼─────────────┤');
-  console.log('│  30.0s      │  27.5s       │  22.48s      │  ~15s       │');
-  console.log('│  100%       │  91.7%       │  74.9%       │  ~50%       │');
-  console.log('└─────────────┴──────────────┴──────────────┴─────────────┘');
-  
+  console.log("┌─────────────┬──────────────┬──────────────┬─────────────┐");
+  console.log("│   Baseline  │   Phase 1.1  │   Phase 1.3  │  Phase 2    │");
+  console.log("│   (Start)   │  (Complete)  │  (Complete)  │  (Planned)  │");
+  console.log("├─────────────┼──────────────┼──────────────┼─────────────┤");
+  console.log("│  30.0s      │  27.5s       │  22.48s      │  ~15s       │");
+  console.log("│  100%       │  91.7%       │  74.9%       │  ~50%       │");
+  console.log("└─────────────┴──────────────┴──────────────┴─────────────┘");
+
   // Infrastructure Reduction
   console.log(`\n${colors.bright}Infrastructure Reduction:${colors.reset}`);
-  console.log('┌──────────────────────────┬────────────┬──────────────┐');
-  console.log('│ Component                │   Lines    │    Status    │');
-  console.log('├──────────────────────────┼────────────┼──────────────┤');
-  console.log(`│ TestEnvironmentManager   │    721     │ ${colors.green}✓ ELIMINATED${colors.reset} │`);
-  console.log(`│ TestSingletonManager     │    518     │ ${colors.green}✓ ELIMINATED${colors.reset} │`);
-  console.log(`│ TestMockManager          │    869     │ ${colors.yellow}⏳ PENDING${colors.reset}   │`);
-  console.log(`│ Database Utilities       │   1,017    │ ${colors.yellow}⏳ PENDING${colors.reset}   │`);
-  console.log('├──────────────────────────┼────────────┼──────────────┤');
-  console.log(`│ ${colors.bright}Total Eliminated${colors.reset}         │ ${colors.green}1,239${colors.reset}      │ ${colors.green}88% done${colors.reset}     │`);
-  console.log(`│ ${colors.bright}Remaining${colors.reset}                │ ${colors.yellow}1,886${colors.reset}      │ ${colors.yellow}12% left${colors.reset}     │`);
-  console.log('└──────────────────────────┴────────────┴──────────────┘');
-  
+  console.log("┌──────────────────────────┬────────────┬──────────────┐");
+  console.log("│ Component                │   Lines    │    Status    │");
+  console.log("├──────────────────────────┼────────────┼──────────────┤");
+  console.log(
+    `│ TestEnvironmentManager   │    721     │ ${colors.green}✓ ELIMINATED${colors.reset} │`,
+  );
+  console.log(
+    `│ TestSingletonManager     │    518     │ ${colors.green}✓ ELIMINATED${colors.reset} │`,
+  );
+  console.log(
+    `│ TestMockManager          │    869     │ ${colors.yellow}⏳ PENDING${colors.reset}   │`,
+  );
+  console.log(
+    `│ Database Utilities       │   1,017    │ ${colors.yellow}⏳ PENDING${colors.reset}   │`,
+  );
+  console.log("├──────────────────────────┼────────────┼──────────────┤");
+  console.log(
+    `│ ${colors.bright}Total Eliminated${colors.reset}         │ ${colors.green}1,239${colors.reset}      │ ${colors.green}88% done${colors.reset}     │`,
+  );
+  console.log(
+    `│ ${colors.bright}Remaining${colors.reset}                │ ${colors.yellow}1,886${colors.reset}      │ ${colors.yellow}12% left${colors.reset}     │`,
+  );
+  console.log("└──────────────────────────┴────────────┴──────────────┘");
+
   // Performance Gains Bar Chart
-  console.log(`\n${colors.bright}Performance Improvement Progress:${colors.reset}`);
-  console.log('');
-  console.log('Baseline  │████████████████████████████████│ 30.0s (100%)');
-  console.log('Phase 1.1 │███████████████████████████▌    │ 27.5s (91.7%)');
-  console.log('Phase 1.3 │████████████████████▌           │ 22.48s (74.9%)');
-  console.log('Phase 2*  │███████████████                 │ ~15s (50%)');
-  console.log('');
-  console.log('          0s    5s    10s   15s   20s   25s   30s');
-  console.log(`          ${colors.green}← Better${colors.reset}                    ${colors.red}Worse →${colors.reset}`);
-  console.log('');
-  console.log('* Projected based on current trends');
-  
+  console.log(
+    `\n${colors.bright}Performance Improvement Progress:${colors.reset}`,
+  );
+  console.log("");
+  console.log("Baseline  │████████████████████████████████│ 30.0s (100%)");
+  console.log("Phase 1.1 │███████████████████████████▌    │ 27.5s (91.7%)");
+  console.log("Phase 1.3 │████████████████████▌           │ 22.48s (74.9%)");
+  console.log("Phase 2*  │███████████████                 │ ~15s (50%)");
+  console.log("");
+  console.log("          0s    5s    10s   15s   20s   25s   30s");
+  console.log(
+    `          ${colors.green}← Better${colors.reset}                    ${colors.red}Worse →${colors.reset}`,
+  );
+  console.log("");
+  console.log("* Projected based on current trends");
+
   // Memory and Complexity Impact
   console.log(`\n${colors.bright}Complexity Reduction Impact:${colors.reset}`);
-  console.log('┌────────────────────┬──────────┬──────────┬────────────┐');
-  console.log('│ Metric             │ Baseline │ Current  │ Reduction  │');
-  console.log('├────────────────────┼──────────┼──────────┼────────────┤');
-  console.log(`│ Manager Classes    │    3     │    0     │   ${colors.green}100%${colors.reset}     │`);
-  console.log(`│ Complexity Score   │   HIGH   │   LOW    │   ${colors.green}71%${colors.reset}      │`);
-  console.log(`│ Test Execution     │  30.0s   │  22.48s  │   ${colors.green}25.1%${colors.reset}    │`);
-  console.log(`│ Infrastructure     │  3,125   │   518*   │   ${colors.green}88%${colors.reset}      │`);
-  console.log('└────────────────────┴──────────┴──────────┴────────────┘');
-  console.log('* Lines of complex test infrastructure code');
-  
+  console.log("┌────────────────────┬──────────┬──────────┬────────────┐");
+  console.log("│ Metric             │ Baseline │ Current  │ Reduction  │");
+  console.log("├────────────────────┼──────────┼──────────┼────────────┤");
+  console.log(
+    `│ Manager Classes    │    3     │    0     │   ${colors.green}100%${colors.reset}     │`,
+  );
+  console.log(
+    `│ Complexity Score   │   HIGH   │   LOW    │   ${colors.green}71%${colors.reset}      │`,
+  );
+  console.log(
+    `│ Test Execution     │  30.0s   │  22.48s  │   ${colors.green}25.1%${colors.reset}    │`,
+  );
+  console.log(
+    `│ Infrastructure     │  3,125   │   518*   │   ${colors.green}88%${colors.reset}      │`,
+  );
+  console.log("└────────────────────┴──────────┴──────────┴────────────┘");
+  console.log("* Lines of complex test infrastructure code");
+
   // Success Metrics
   console.log(`\n${colors.bright}Phase 1.3 Success Metrics:${colors.reset}`);
   const metrics = [
-    ['TestSingletonManager eliminated', true],
-    ['518 lines of code removed', true],
-    ['25.1% performance improvement', true],
-    ['Memory usage reduced', true],
-    ['88% reduction maintained', true],
-    ['Test reliability improved', true]
+    ["TestSingletonManager eliminated", true],
+    ["518 lines of code removed", true],
+    ["25.1% performance improvement", true],
+    ["Memory usage reduced", true],
+    ["88% reduction maintained", true],
+    ["Test reliability improved", true],
   ];
-  
+
   metrics.forEach(([metric, achieved]) => {
-    const status = achieved ? `${colors.green}✓${colors.reset}` : `${colors.red}✗${colors.reset}`;
+    const status = achieved
+      ? `${colors.green}✓${colors.reset}`
+      : `${colors.red}✗${colors.reset}`;
     console.log(`  ${status} ${metric}`);
   });
-  
+
   // Next Phase Preview
-  console.log(`\n${colors.bright}${colors.cyan}NEXT PHASE PREVIEW - Phase 2${colors.reset}`);
-  console.log('┌──────────────────────────────────────────────────────┐');
-  console.log('│ Target: TestMockManager (869 lines)                  │');
-  console.log('│ Strategy: Replace with direct Vitest mocking         │');
-  console.log('│ Expected Performance Gain: ~30% additional           │');
-  console.log('│ Expected Total Reduction: 95% of infrastructure      │');
-  console.log('│ Estimated Completion: 2-3 days                       │');
-  console.log('└──────────────────────────────────────────────────────┘');
-  
-  console.log(`\n${colors.bright}${colors.green}Phase 1.3 Complete - All objectives achieved!${colors.reset}\n`);
+  console.log(
+    `\n${colors.bright}${colors.cyan}NEXT PHASE PREVIEW - Phase 2${colors.reset}`,
+  );
+  console.log("┌──────────────────────────────────────────────────────┐");
+  console.log("│ Target: TestMockManager (869 lines)                  │");
+  console.log("│ Strategy: Replace with direct Vitest mocking         │");
+  console.log("│ Expected Performance Gain: ~30% additional           │");
+  console.log("│ Expected Total Reduction: 95% of infrastructure      │");
+  console.log("│ Estimated Completion: 2-3 days                       │");
+  console.log("└──────────────────────────────────────────────────────┘");
+
+  console.log(
+    `\n${colors.bright}${colors.green}Phase 1.3 Complete - All objectives achieved!${colors.reset}\n`,
+  );
 }
 
 generateChart();