… docs/performance-reports/phase-1-3-metrics.json
… docs/performance-reports/phase-1-3-performance-analysis.md
… migration-backups/tests_config_enhanced-test-setup.js.backup
… migration-backups/tests_setup-vitest.js.backup
… migration-backups/tests_unit_complete-isolation-demo.test.js.backup
… migration-backups/tests_unit_database-client.test.js.backup
… migration-backups/tests_unit_database-environment.test.js.backup
… migration-backups/tests_unit_database-singleton.test.js.backup
… migration-backups/tests_unit_simple-helpers.test.js.backup
… migration-backups/tests_unit_test-environment-manager-usage-examples.test.js.backup
… migration-backups/tests_unit_test-environment-manager.test.js.backup
… migration-backups/tests_utils_integration-test-patterns.js.backup
… migration-backups/tests_utils_integration-test-strategy.js.backup
… migration-backups/tests_utils_test-environment-manager.js.backup
… playwright-report/index.html
… tests/config/performance-isolation-config.js
… tests/e2e/global-setup.js
… tests/helpers/simple-helpers.js
… tests/unit/test-singleton-manager.test.js
… tests/utils/test-singleton-manager.js